Rename MapD -> OmnisciDB
