+++ rbc/tests/test_omnisci_udtf.py
@@ -1,12 +1,12 @@
 import pytest
-rbc_omnisci = pytest.importorskip('rbc.mapd')
+rbc_omnisci = pytest.importorskip('rbc.omniscidb')
 
 
 def omnisci_is_available():
     """Return True if OmniSci server is accessible.
     """
     config = rbc_omnisci.get_client_config()
-    omnisci = rbc_omnisci.RemoteMapD(**config)
+    omnisci = rbc_omnisci.RemoteOmnisci(**config)
     client = omnisci.client
     try:
         version = client(MapD=dict(get_version=()))['MapD']['get_version']
@@ -24,7 +24,7 @@ pytestmark = pytest.mark.skipif(not is_available, reason=reason)
 @pytest.fixture(scope='module')
 def omnisci():
     config = rbc_omnisci.get_client_config(debug=not True)
-    m = rbc_omnisci.RemoteMapD(**config)
+    m = rbc_omnisci.RemoteOmnisci(**config)
     table_name = 'rbc_test_omnisci_udtf'
     m.sql_execute('DROP TABLE IF EXISTS {table_name}'.format(**locals()))
